requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0
pytest>=7.0.0
pytest-cov>=4.0.0
